Take out card row into it's own component
